modification des img des posts

+++ controler/backend/post_controler.php
@@ -14,11 +14,22 @@ function editPost($postId){
 function deletePost($id,$chapter){
     $postManager = new JeanForteroche\Blog\Model\PostManager();
     $postManager->deletePost($id);
-    unlink('web/img/portfolio/thumbnails/'.$chapter.'.jpg');   
+    if(file_exists('web/img/portfolio/thumbnails/'.$chapter.'.jpg')){
+        unlink('web/img/portfolio/thumbnails/'.$chapter.'.jpg');   
+    }
+    if(file_exists('web/img/portfolio/fullsize/'.$chapter.'.jpg')){
+        unlink('web/img/portfolio/fullsize/'.$chapter.'.jpg');   
+    }
     header('Location: index.php');
 }
 
-function updatePost(){ 
+function updatePost($title,$content,$chapter,$imgFiles){ 
+    $FileManager = new JeanForteroche\Blog\Model\FileManager();
+    $FileManager->upload($imgFiles,$chapter);
+    
+    $img=pathinfo($imgFiles['img']['name']);
+    $imgUrl = 'web/img/portfolio/thumbnails/'.$chapter.'.'.$img['extension'];
+    
     $postManager = new JeanForteroche\Blog\Model\PostManager();
     $postManager->updatePost($_POST['title'],$_POST['content'],$_GET['id']);   
     header('Location: index.php');
